+++ lux/extensions/odm/views.py
@@ -1,7 +1,5 @@
 from sqlalchemy.exc import DataError
 from sqlalchemy.orm.exc import NoResultFound
-from sqlalchemy.orm import load_only
-from sqlalchemy import desc
 
 from pulsar import PermissionDenied, MethodNotAllowed, Http404
 from pulsar.apps.wsgi import Json
@@ -22,36 +20,15 @@ class RestRouter(rest.RestRouter):
     def urlargs(self, request):
         return request.urlargs
 
-    def query(self, request, session, *filters):
-        """
-        Returns a query object for the model.
-
-        The loading of columns the user does not have read
-        access to is deferred. This is only a performance enhancement.
-
-        :param request:     request object
-        :param session:     SQLAlchemy session
-        :return:            query object
-        """
-        entities = self.columns_with_permission(request, rest.READ)
-        if not entities:
-            raise PermissionDenied
-        model = self.model(request.app)
-        db_model = model.db_model()
-        db_columns = model.db_columns(self.column_fields(entities))
-        query = session.query(db_model).options(load_only(*db_columns))
-        if filters:
-            query = query.filter(*filters)
-        return model.query(query)
-
     # RestView implementation
     def get_instance(self, request, session=None, **args):
         odm = request.app.odm()
         args = args or self.urlargs(request)
         if not args:  # pragma    nocover
             raise Http404
+        model = self.model(request)
         with odm.begin(session=session) as session:
-            query = self.query(request, session)
+            query = model.query(request, session)
             try:
                 return query.filter_by(**args).one()
             except (DataError, NoResultFound):
@@ -89,54 +66,6 @@ class RestRouter(rest.RestRouter):
         with odm.begin(session=session) as session:
             session.delete(instance)
 
-    def serialise_model(self, request, data, **kw):
-        """
-        Makes a model instance JSON-friendly. Removes fields that the
-        user does not have read access to.
-
-        :param request:     request object
-        :param data:        data
-        :param kw:          not used
-        :return:            dict
-        """
-        exclude = self.columns_without_permission(request, rest.READ)
-        exclude = self.column_fields(exclude, 'name')
-        model = self.model(request.app)
-        return model.tojson(request, data, exclude=exclude)
-
-    def meta(self, request):
-        meta = super().meta(request)
-        odm = request.app.odm()
-        with odm.begin() as session:
-            query = self.query(request, session)
-            meta['total'] = query.count()
-        return meta
-
-    def _do_filter(self, request, model, query, field, op, value):
-        if value == '':
-            value = None
-        odm = request.app.odm()
-        field = getattr(odm[model.name], field)
-        if op == 'eq':
-            query = query.filter(field == value)
-        elif op == 'gt':
-            query = query.filter(field > value)
-        elif op == 'ge':
-            query = query.filter(field >= value)
-        elif op == 'lt':
-            query = query.filter(field < value)
-        elif op == 'le':
-            query = query.filter(field <= value)
-        return query
-
-    def _do_sortby(self, request, query, entry, direction):
-        columns = self.model(request).db_columns()
-        if entry in columns:
-            if direction == 'desc':
-                entry = desc(entry)
-            return query.order_by(entry)
-        return query
-
 
 class CRUD(RestRouter):
     '''A Router for handling CRUD JSON requests for a database model
@@ -153,7 +82,7 @@ class CRUD(RestRouter):
         self.check_model_permission(request, rest.READ)
         # Columns the user doesn't have access to are dropped by
         # serialise_model
-        return self.collection_response(request)
+        return self.model(request).collection_response(request)
 
     def post(self, request):
         '''Create a new model
@@ -163,8 +92,8 @@ class CRUD(RestRouter):
             raise MethodNotAllowed
 
         self.check_model_permission(request, rest.CREATE)
-        columns = self.columns_with_permission(request, rest.CREATE)
-        columns = self.column_fields(columns, 'name')
+        columns = model.columns_with_permission(request, rest.CREATE)
+        columns = model.column_fields(columns, 'name')
 
         data, files = request.data_and_files()
         form = model.form(request, data=data, files=files)
@@ -181,7 +110,7 @@ class CRUD(RestRouter):
                 form.add_error_message(str(exc))
                 data = form.tojson()
             else:
-                data = self.serialise(request, instance)
+                data = model.serialise(request, instance)
                 request.response.status_code = 201
         else:
             data = form.tojson()
@@ -201,7 +130,7 @@ class CRUD(RestRouter):
         model = self.model(request.app)
 
         if backend.has_permission(request, model.name, rest.READ):
-            meta = self.meta(request)
+            meta = model.meta(request)
             return Json(meta).http_response(request)
         raise PermissionDenied
 
@@ -218,7 +147,7 @@ class CRUD(RestRouter):
 
             if request.method == 'GET':
                 self.check_model_permission(request, rest.READ)
-                data = self.serialise(request, instance)
+                data = model.serialise(request, instance)
 
             elif request.method == 'HEAD':
                 self.check_model_permission(request, rest.READ)
@@ -231,8 +160,8 @@ class CRUD(RestRouter):
                     raise MethodNotAllowed
 
                 self.check_model_permission(request, rest.UPDATE)
-                columns = self.columns_with_permission(request, rest.UPDATE)
-                columns = self.column_fields(columns, 'name')
+                columns = model.columns_with_permission(request, rest.UPDATE)
+                columns = model.column_fields(columns, 'name')
 
                 data, files = request.data_and_files()
                 form = form_class(request, data=data, files=files,
@@ -246,7 +175,7 @@ class CRUD(RestRouter):
 
                     instance = self.update_model(request, instance,
                                                  filtered_data)
-                    data = self.serialise(request, instance)
+                    data = model.serialise(request, instance)
                 else:
                     data = form.tojson()
